feat: Agregar funcionalidades de suscripción con CRUD completo en la base de datos
Nota: se separanfuncionalidades para evitar codigo spagetti . Quedan pendiente implementar las funcionalidades en el DOM
el dom queda preconfigurado para manejar la funcionalidad del servidor
