+++ php/quejasSugerencias.php
@@ -202,129 +202,6 @@ function deleteMessageOnDataBase($conn, $id_comentario) {
     }
 }
 
-// ================================= SUSCRIPCION =================================
-
-// funcion para guardar el correo en la base de datos
-function saveEmailUserSuscriptionOnDataBase($conn, $email) {
-    // Validación de correo
-    if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
-        return "Correo no válido.";
-    }
-
-    // Preparamos la consulta para evitar inyección SQL
-    $query = "INSERT INTO SUSCRIPCIONES (CORREO, ESTADO) VALUES (?, 'Suscrito')";
-    $stmt = $conn->prepare($query);
-    
-    if ($stmt === false) {
-        return "Error en la preparación de la consulta: " . $conn->error;
-    }
-
-    // Vinculamos el parámetro del correo
-    $stmt->bind_param("s", $email);
-
-    // Ejecutamos la consulta
-    if ($stmt->execute()) {
-        return "Correo suscrito exitosamente.";
-    } else {
-        return "Error al suscribir el correo: " . $stmt->error;
-    }
-
-    $stmt->close();
-}
-// funcion para editar el correo en la base de datos
-function editEmailUserSuscriptionOnDataBase($conn, $id_correo, $correo, $estado) {
-    // Verificar que se pasen los parámetros correctos
-    if (empty($id_correo) || empty($correo) || empty($estado)) {
-        return "Error: Debes proporcionar el ID, correo y estado para editar.";
-    }
-
-    // Consulta SQL para actualizar el correo y el estado, y establecer la fecha de edición
-    $query = "UPDATE SUSCRIPCIONES 
-            SET CORREO = ?, ESTADO = ?, FECHA_CREACION = CURRENT_TIMESTAMP
-            WHERE ID_SUSCRIPCION = ?";
-    
-    $stmt = $conn->prepare($query);
-    if ($stmt === false) {
-        return "Error en la preparación de la consulta: " . $conn->error;
-    }
-
-    // Vinculamos los parámetros: 's' para string (correo y estado), 'i' para integer (id_correo)
-    $stmt->bind_param("ssi", $correo, $estado, $id_correo);
-
-    // Ejecutamos la consulta
-    if ($stmt->execute()) {
-        return "Suscripción actualizada correctamente.";
-    } else {
-        return "Error al actualizar la suscripción: " . $stmt->error;
-    }
-
-    $stmt->close();
-}
-// funcion para leer el correo en la base de datos
-function readEmailUserSuscriptionOnDataBase($conn, $id_correo = null, $correo = null) {
-    // Si se proporciona un ID o un correo, filtrar según el caso
-    if ($id_correo) {
-        // Filtrar por ID
-        $query = "SELECT * FROM SUSCRIPCIONES WHERE ID_SUSCRIPCION = ?";
-        $stmt = $conn->prepare($query);
-        if ($stmt === false) {
-            return "Error en la preparación de la consulta: " . $conn->error;
-        }
-        $stmt->bind_param("i", $id_correo); // 'i' para entero (ID)
-    } elseif ($correo) {
-        // Filtrar por correo
-        $query = "SELECT * FROM SUSCRIPCIONES WHERE CORREO = ?";
-        $stmt = $conn->prepare($query);
-        if ($stmt === false) {
-            return "Error en la preparación de la consulta: " . $conn->error;
-        }
-        $stmt->bind_param("s", $correo); // 's' para string (correo)
-    } else {
-        // Si no se proporciona ni ID ni correo, obtener todos los registros
-        $query = "SELECT * FROM SUSCRIPCIONES";
-        $stmt = $conn->prepare($query);
-        if ($stmt === false) {
-            return "Error en la preparación de la consulta: " . $conn->error;
-        }
-    }
-
-    // Ejecutamos la consulta
-    $stmt->execute();
-    $resultado = $stmt->get_result();
-    
-    if ($resultado->num_rows > 0) {
-        // Devolvemos los resultados como un array asociativo
-        return $resultado->fetch_all(MYSQLI_ASSOC);
-    } else {
-        return "No se encontraron suscripciones.";
-    }
-
-    $stmt->close();
-}
-// funcion para eliminar el correo en la base de datos
-function deleteEmailUserSuscriptionOnDataBase($conn, $id_correo) {
-    // Preparamos la consulta para eliminar la suscripción
-    $query = "DELETE FROM SUSCRIPCIONES WHERE ID_SUSCRIPCION = ?";
-    $stmt = $conn->prepare($query);
-    
-    if ($stmt === false) {
-        return "Error en la preparación de la consulta: " . $conn->error;
-    }
-
-    // Vinculamos el parámetro del ID
-    $stmt->bind_param("i", $id_correo); // 'i' para entero
-
-    // Ejecutamos la consulta
-    if ($stmt->execute()) {
-        return "Correo eliminado de la base de datos.";
-    } else {
-        return "Error al eliminar el correo: " . $stmt->error;
-    }
-
-    $stmt->close();
-}
-
-
 // ================================= ENRUTAMIENTO =================================
 if (isset($_GET['action'])) {
     $action = $_GET['action'];
@@ -412,70 +289,6 @@ if (isset($_GET['action'])) {
                 }
             } else {
                 $data = ["error" => "ID_COMENTARIO es obligatorio"];
-            }
-            break;
-
-        case 'guardarSuscripcion':
-            if (isset($_POST['correo'])) {
-                $correo = $_POST['correo'];
-                // Validación del correo (opcional)
-                if (!filter_var($correo, FILTER_VALIDATE_EMAIL)) {
-                    $data = ["error" => "Correo electrónico no válido"];
-                } else {
-                    // Llamar a la función para guardar en la base de datos
-                    $result = saveEmailUserSuscriptionOnDataBase($conn, $correo);
-                    if ($result) {
-                        $data = ["success" => "Correo guardado correctamente"];
-                    } else {
-                        $data = ["error" => "Hubo un problema al guardar el correo"];
-                    }
-                }
-            } else {
-                $data = ["error" => "Correo no proporcionado"];
-            }
-            break;
-
-        // Acción para editar el estado de suscripción
-        case 'editarSuscripcion':
-            if (isset($_POST['id_correo']) && isset($_POST['estado'])) {
-                $id_correo = $_POST['id_correo'];
-                $estado = $_POST['estado'];
-                // Llamar a la función para editar el correo en la base de datos
-                $result = editEmailUserSuscriptionOnDataBase($conn, $id_correo, $estado);
-                if ($result) {
-                    $data = ["success" => "Estado de suscripción actualizado"];
-                } else {
-                    $data = ["error" => "Hubo un problema al actualizar el estado"];
-                }
-            } else {
-                $data = ["error" => "Faltan parámetros para editar"];
-            }
-            break;
-
-        // Acción para leer todos los correos de suscripción
-        case 'leerSuscripciones':
-            // Llamar a la función para leer todos los correos de suscripción
-            $suscripciones = readEmailUserSuscriptionOnDataBase($conn);
-            if ($suscripciones) {
-                $data = ["success" => "Suscripciones obtenidas", "suscripciones" => $suscripciones];
-            } else {
-                $data = ["error" => "No se encontraron suscripciones"];
-            }
-            break;
-
-        // Acción para eliminar una suscripción
-        case 'eliminarSuscripcion':
-            if (isset($_POST['id_correo'])) {
-                $id_correo = $_POST['id_correo'];
-                // Llamar a la función para eliminar el correo en la base de datos
-                $result = deleteEmailUserSuscriptionOnDataBase($conn, $id_correo);
-                if ($result) {
-                    $data = ["success" => "Correo eliminado correctamente"];
-                } else {
-                    $data = ["error" => "Hubo un problema al eliminar el correo"];
-                }
-            } else {
-                $data = ["error" => "Faltan parámetros para eliminar"];
             }
             break;
